Add AnnotationPropertyNameCollectionFactoryTest
